test case refactor for deployments, sts, clonesets and asts
