+++ pkg/factory/test_common.go
@@ -1,0 +1,362 @@
+package factory
+
+import (
+	"strconv"
+
+	"github.com/samber/lo"
+	appsv1 "k8s.io/api/apps/v1"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/equality"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/rand"
+	"k8s.io/apimachinery/pkg/util/uuid"
+	"k8s.io/utils/pointer"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	kruiseappsv1alpha1 "github.com/openkruise/kruise-api/apps/v1alpha1"
+	kruiseappsv1beta1 "github.com/openkruise/kruise-api/apps/v1beta1"
+	risingwavev1alpha1 "github.com/risingwavelabs/risingwave-operator/apis/risingwave/v1alpha1"
+	"github.com/risingwavelabs/risingwave-operator/pkg/consts"
+)
+
+type Testcase[T any, K any] struct {
+	name                    string
+	risingwave              *risingwavev1alpha1.RisingWave
+	component               string
+	podTemplate             map[string]risingwavev1alpha1.RisingWavePodTemplate
+	group                   T
+	expectedUpgradeStrategy K
+	restartAt               *metav1.Time
+}
+
+type deploymentTestcase Testcase[risingwavev1alpha1.RisingWaveComponentGroup, *appsv1.DeploymentStrategy]
+type stsTestcase Testcase[risingwavev1alpha1.RisingWaveComputeGroup, *appsv1.StatefulSetUpdateStrategy]
+type clonesetTestcase Testcase[risingwavev1alpha1.RisingWaveComponentGroup, *kruiseappsv1alpha1.CloneSetUpdateStrategy]
+type advancedSTSTestcase Testcase[risingwavev1alpha1.RisingWaveComputeGroup, *kruiseappsv1beta1.StatefulSetUpdateStrategy]
+
+// type advancedSTSTestcase Testcase[risingwavev1alpha1.RisingWaveComponentGroup, *kruiseappsv1alpha1.CloneSetUpdateStrategy]
+
+func MapContains[K, V comparable](a, b map[K]V) bool {
+	if len(a) < len(b) {
+		return false
+	}
+
+	for k, v := range b {
+		va, ok := a[k]
+		if !ok || va != v {
+			return false
+		}
+	}
+
+	return true
+}
+
+func MapContainsWith[K comparable, V any](a, b map[K]V, equals func(a, b V) bool) bool {
+	if len(a) < len(b) {
+		return false
+	}
+
+	for k, v := range b {
+		va, ok := a[k]
+		if !ok || !equals(va, v) {
+			return false
+		}
+	}
+
+	return true
+}
+
+func MapEquals[K, V comparable](a, b map[K]V) bool {
+	if len(a) == 0 && len(b) == 0 {
+		return true
+	} else if len(a) == 0 || len(b) == 0 || len(a) != len(b) {
+		return false
+	} else {
+		for k, v := range a {
+			vb, ok := b[k]
+			if !ok || v != vb {
+				return false
+			}
+		}
+		return true
+	}
+}
+
+func HasLabels[T client.Object](obj T, labels map[string]string, exact bool) bool {
+	for k, v := range labels {
+		v1, ok := obj.GetLabels()[k]
+		if !ok || v != v1 {
+			return false
+		}
+	}
+	if exact && len(obj.GetLabels()) != len(labels) {
+		return false
+	}
+	return true
+}
+
+func HasAnnotations[T client.Object](obj T, annotations map[string]string, exact bool) bool {
+	for k, v := range annotations {
+		v1, ok := obj.GetAnnotations()[k]
+		if !ok || v != v1 {
+			return false
+		}
+	}
+	if exact && len(obj.GetAnnotations()) != len(annotations) {
+		return false
+	}
+	return true
+}
+
+func IsServiceType(svc *corev1.Service, t corev1.ServiceType) bool {
+	return svc.Spec.Type == t
+}
+
+func HasTCPServicePorts(svc *corev1.Service, ports map[string]int32) bool {
+	svcPorts := make(map[string]corev1.ServicePort)
+	for _, port := range svc.Spec.Ports {
+		svcPorts[port.Name] = port
+	}
+
+	for name, port := range ports {
+		svcPort, ok := svcPorts[name]
+		if !ok || (svcPort.Protocol != corev1.ProtocolTCP && svcPort.Protocol != "") || svcPort.Port != port {
+			return false
+		}
+	}
+
+	return true
+}
+
+func HasServiceSelector(svc *corev1.Service, selector map[string]string) bool {
+	return equality.Semantic.DeepEqual(svc.Spec.Selector, selector)
+}
+
+func ComponentLabels(risingwave *risingwavev1alpha1.RisingWave, component string, sync bool) map[string]string {
+	labels := map[string]string{
+		consts.LabelRisingWaveName:      risingwave.Name,
+		consts.LabelRisingWaveComponent: component,
+	}
+	if sync {
+		labels[consts.LabelRisingWaveGeneration] = strconv.FormatInt(risingwave.Generation, 10)
+	} else {
+		labels[consts.LabelRisingWaveGeneration] = consts.NoSync
+	}
+
+	if component == consts.ComponentFrontend {
+		labels = mergeMap(labels, risingwave.Spec.Global.ServiceMeta.Labels)
+	}
+
+	return labels
+}
+
+func ComponentGroupLabels(risingwave *risingwavev1alpha1.RisingWave, component string, group *string, sync bool) map[string]string {
+	labels := map[string]string{
+		consts.LabelRisingWaveName:      risingwave.Name,
+		consts.LabelRisingWaveComponent: component,
+	}
+	if sync {
+		labels[consts.LabelRisingWaveGeneration] = strconv.FormatInt(risingwave.Generation, 10)
+	} else {
+		labels[consts.LabelRisingWaveGeneration] = consts.NoSync
+	}
+
+	if group != nil {
+		labels[consts.LabelRisingWaveGroup] = *group
+		labels = mergeMap(labels, risingwave.Spec.Global.Metadata.Labels)
+	}
+
+	return labels
+}
+
+func ComponentAnnotations(risingwave *risingwavev1alpha1.RisingWave, component string) map[string]string {
+	annotations := map[string]string{}
+	if component == consts.ComponentFrontend {
+		annotations = mergeMap(annotations, risingwave.Spec.Global.ServiceMeta.Annotations)
+	}
+
+	return annotations
+}
+
+func ComponentGroupAnnotations(risingwave *risingwavev1alpha1.RisingWave, group *string) map[string]string {
+	annotations := map[string]string{}
+	if group != nil {
+		annotations = mergeMap(annotations, risingwave.Spec.Global.Metadata.Annotations)
+	}
+
+	return annotations
+}
+
+func PodSelector(risingwave *risingwavev1alpha1.RisingWave, component string, group *string) map[string]string {
+	labels := map[string]string{
+		consts.LabelRisingWaveName:      risingwave.Name,
+		consts.LabelRisingWaveComponent: component,
+	}
+	if group != nil {
+		labels[consts.LabelRisingWaveGroup] = *group
+	}
+	return labels
+}
+
+func ControlledBy(owner, ownee client.Object) bool {
+	controllerRef, ok := lo.Find(ownee.GetOwnerReferences(), func(ref metav1.OwnerReference) bool {
+		return ref.Controller != nil && *ref.Controller
+	})
+	if !ok {
+		return false
+	}
+	return controllerRef.UID == owner.GetUID()
+}
+
+func NewTestRisingwave(patches ...func(r *risingwavev1alpha1.RisingWave)) *risingwavev1alpha1.RisingWave {
+	r := &risingwavev1alpha1.RisingWave{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:       "test",
+			Namespace:  rand.String(10),
+			Generation: int64(rand.Int()),
+			UID:        uuid.NewUUID(),
+		},
+	}
+	for _, patch := range patches {
+		patch(r)
+	}
+	return r
+}
+
+func SetupMetaPorts(r *risingwavev1alpha1.RisingWave, ports map[string]int32) {
+	r.Spec.Components.Meta.Ports = risingwavev1alpha1.RisingWaveComponentMetaPorts{
+		RisingWaveComponentCommonPorts: risingwavev1alpha1.RisingWaveComponentCommonPorts{
+			ServicePort: ports[consts.PortService],
+			MetricsPort: ports[consts.PortMetrics],
+		},
+		DashboardPort: ports[consts.PortDashboard],
+	}
+}
+
+func SetupFrontendPorts(r *risingwavev1alpha1.RisingWave, ports map[string]int32) {
+	r.Spec.Components.Frontend.Ports = risingwavev1alpha1.RisingWaveComponentCommonPorts{
+		ServicePort: ports[consts.PortService],
+		MetricsPort: ports[consts.PortMetrics],
+	}
+}
+
+func SetupComputePorts(r *risingwavev1alpha1.RisingWave, ports map[string]int32) {
+	r.Spec.Components.Compute.Ports = risingwavev1alpha1.RisingWaveComponentCommonPorts{
+		ServicePort: ports[consts.PortService],
+		MetricsPort: ports[consts.PortMetrics],
+	}
+}
+
+func SetupCompactorPorts(r *risingwavev1alpha1.RisingWave, ports map[string]int32) {
+	r.Spec.Components.Compactor.Ports = risingwavev1alpha1.RisingWaveComponentCommonPorts{
+		ServicePort: ports[consts.PortService],
+		MetricsPort: ports[consts.PortMetrics],
+	}
+}
+
+func SetupConnectorPorts(r *risingwavev1alpha1.RisingWave, ports map[string]int32) {
+	r.Spec.Components.Connector.Ports = risingwavev1alpha1.RisingWaveComponentCommonPorts{
+		ServicePort: ports[consts.PortService],
+		MetricsPort: ports[consts.PortMetrics],
+	}
+}
+
+func deepEqual[T any](x, y T) bool {
+	return equality.Semantic.DeepEqual(x, y)
+}
+
+func listContains[T comparable](a, b []T) bool {
+	if len(b) > len(a) {
+		return false
+	}
+
+	m := make(map[T]int)
+	for _, x := range a {
+		m[x]++
+	}
+
+	for _, x := range b {
+		c := m[x]
+		if c == 0 {
+			return false
+		}
+		m[x]--
+	}
+	return true
+}
+
+func listContainsByKey[T any, K comparable](a, b []T, key func(*T) K, equals func(x, y T) bool) bool {
+	aKeys, bKeys := make(map[K]T), make(map[K]T)
+	for i, x := range a {
+		aKeys[key(&x)] = a[i]
+	}
+	for i, x := range b {
+		bKeys[key(&x)] = b[i]
+	}
+
+	return MapContainsWith(aKeys, bKeys, equals)
+}
+
+func MatchesPodTemplate(podSpec *corev1.PodTemplateSpec, podTemplate *risingwavev1alpha1.RisingWavePodTemplateSpec) bool {
+	if podTemplate == nil {
+		return true
+	}
+
+	if !(MapContains(podSpec.Labels, podTemplate.Labels) &&
+		MapContains(podSpec.Annotations, podTemplate.Annotations)) {
+		return false
+	}
+
+	pSpec, tSpec := podSpec.Spec, podTemplate.Spec
+	pSpec.Containers = pSpec.Containers[1:]
+	tSpec.Containers = tSpec.Containers[1:]
+
+	// Trick: remove the readinessGate to pass the match.
+	pSpec.ReadinessGates = nil
+	tSpec.ReadinessGates = nil
+
+	// Check volumes first.
+	if !listContainsByKey(pSpec.Volumes, tSpec.Volumes, func(x *corev1.Volume) string { return x.Name }, deepEqual[corev1.Volume]) {
+		return false
+	}
+	pSpec.Volumes, tSpec.Volumes = nil, nil
+
+	// Set default enable service links to false.
+	if tSpec.EnableServiceLinks == nil {
+		tSpec.EnableServiceLinks = pointer.Bool(false)
+	}
+
+	if !equality.Semantic.DeepEqual(pSpec, tSpec) {
+		return false
+	}
+
+	pContainer, tContainer := podSpec.Spec.Containers[0], podTemplate.Spec.Containers[0]
+
+	// Only check the
+	//   * SecurityContext
+	//   * Env
+	//   * EnvFrom
+	//   * VolumeDevices
+	return deepEqual(pContainer.SecurityContext, tContainer.SecurityContext) &&
+		listContainsByKey(pContainer.VolumeMounts, tContainer.VolumeMounts, func(t *corev1.VolumeMount) string { return t.MountPath }, deepEqual[corev1.VolumeMount]) &&
+		listContainsByKey(pContainer.Env, tContainer.Env, func(t *corev1.EnvVar) string { return t.Name }, deepEqual[corev1.EnvVar]) &&
+		listContainsByKey(pContainer.EnvFrom, tContainer.EnvFrom, func(t *corev1.EnvFromSource) string { return t.Prefix }, deepEqual[corev1.EnvFromSource])
+}
+
+func NewPodTemplate(patches ...func(t *risingwavev1alpha1.RisingWavePodTemplateSpec)) *risingwavev1alpha1.RisingWavePodTemplateSpec {
+	t := &risingwavev1alpha1.RisingWavePodTemplateSpec{
+		Spec: corev1.PodSpec{
+			Containers: []corev1.Container{
+				{
+					Name: "",
+				},
+			},
+		},
+	}
+
+	for _, patch := range patches {
+		patch(t)
+	}
+
+	return t
+}
